PO-2255 - code review updates

+++ src/integrationTest/java/uk/gov/hmcts/opal/service/GenericReportServiceTest.java
@@ -164,7 +164,7 @@ class GenericReportServiceTest extends AbstractIntegrationTest {
         }
 
         @Override
-        public int getNumberOfRecords() {
+        public long getNumberOfRecords() {
             return numberOfRecords;
         }
